Fix variable type of block sample rate.

+++ src/mdsim/correlation.hpp
@@ -93,7 +93,7 @@ public:
     /** set total simulation time */
     void time(float const& value, float const& timestep);
     /** set sample rate for lowest block level */
-    void sample_rate(float const& sample_rate);
+    void sample_rate(unsigned int const& sample_rate);
     /** set block size */
     void block_size(unsigned int const& value);
     /** set maximum number of samples per block */
@@ -106,7 +106,7 @@ public:
     /** returns total simulation time */
     float const& time() const { return m_time; }
     /** returns sample rate for lowest block level */
-    float const& sample_rate() const { return m_sample_rate; }
+    unsigned int const& sample_rate() const { return m_sample_rate; }
     /** returns block size */
     unsigned int const& block_size() const { return m_block_size; }
     /** returns block shift */
@@ -148,7 +148,7 @@ private:
     /** simulation timestep */
     float m_timestep;
     /** sample rate for lowest block level */
-    float m_sample_rate;
+    unsigned int m_sample_rate;
     /** total number of simulation steps */
     uint64_t m_steps;
     /** total simulation time */
